chore: refactor NewRootCmd into separate pkg (#535)
