+++ cmd/celestia-appd/cmd/root.go
@@ -1,4 +1,4 @@
-package main
+package cmd
 
 import (
 	"fmt"
@@ -36,6 +36,8 @@ import (
 	dbm "github.com/tendermint/tm-db"
 )
 
+const EnvPrefix = "CELESTIA"
+
 // NewRootCmd creates a new root command for celestia-appd. It is called once in the
 // main function.
 func NewRootCmd() *cobra.Command {
@@ -56,7 +58,7 @@ func NewRootCmd() *cobra.Command {
 		WithAccountRetriever(types.AccountRetriever{}).
 		WithBroadcastMode(flags.BroadcastBlock).
 		WithHomeDir(app.DefaultNodeHome).
-		WithViper(envPrefix)
+		WithViper(EnvPrefix)
 
 	rootCmd := &cobra.Command{
 		Use:   "celestia-appd",
